requirements.txt
django==4.2.11
flake8==6.1.0
python-dotenv==1.0.1
psycopg2-binary